requirements.txt
certifi==2018.8.24
chardet==3.0.4
Click==7.0
click-completion==0.4.1
colorama==0.3.9
idna==2.7
Jinja2==2.10
MarkupSafe==1.0
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
saltant-py==0.1.0
shellingham==1.2.6
six==1.11.0
tabulate==0.8.2
urllib3==1.23
